% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImmunoSeq.R
\name{readImmunoSeq}
\alias{readImmunoSeq}
\alias{getStandard}
\alias{readFiles}
\title{Read ImmunoSeq files}
\usage{
readImmunoSeq(path, recursive = FALSE)

getStandard(clone_file, airr_fields)

readFiles(clone_file, progress)
}
\arguments{
\item{path}{Path to the directory containing tab-delimited files.  Only
files with the extension .tsv are imported.  The names of the data frames are
the same as names of the files.}

\item{clone_file}{.tsv file containing results from AIRRSeq pipeline}

\item{airr_fields}{A character vector of MiAIRR headers}
}
\value{
Returns a tibble with MiAIRR headers and repertoire_id

Tibble of given data with MiAIRR fields

Tibble in MiAIRR format
}
\description{
Imports tab-separated value (.tsv) files exported by the Adaptive
Biotechnologies ImmunoSEQ analyzer, BGI IR-SEQ, MiXCR and stores
them as MiAIRR compliant tibble.
}
\details{
May import tab-delimited files containing antigen receptor
sequencing from with the following header set.\tabular{lllllll}{
   MiAIRR field ID \tab Type1 \tab Type2 \tab Type3 \tab Type4 \tab Type5 \tab Type6 \cr
   junction \tab nucleotide \tab nucleotide \tab nucleotide \tab nucleotide.CDR3.in.lowercase . \tab nucleotide( CDR3   in   lowercase ) \tab nSeqCDR3 \cr
   junction_aa \tab aminoAcid \tab aminoAcid \tab aminoAcid \tab aminoAcid.CDR3.in.lowercase . \tab aminoAcid( CDR3   in   lowercase ) \tab aaSeqCDR3 \cr
   duplicate_count \tab count ( reads ) \tab count ( templates / reads ) \tab count ( templates ) \tab cloneCount \tab cloneCount \tab cloneCount \cr
   v_call \tab vGeneName \tab vGeneName \tab vGeneName \tab vGene \tab vGene \tab allVHitsWithScore \cr
   d_call \tab dGeneName \tab dGeneName \tab dGeneName \tab dGene \tab dGene \tab allDHitsWithScore \cr
   j_call \tab jGeneName \tab jGeneName \tab jGeneName \tab jGene \tab jGene \tab allJHitsWithScore \cr
}
}
\section{Get standard headers}{


Retrives MiAIRR standard compliant fields from the clone files
}

\section{Read clone file from path}{


Given the path to a single AIRRSeq clone file, generate MiAIRR compliant tibble.
}

\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeqTest")

study_table <- readImmunoSeq(path = file.path,
                             recursive = FALSE)
}
